requirements.txt
starlette
python-jose
fastapi
uvicorn[standard]
sqlalchemy
pydantic
alembic
python-dotenv
mysql-connector-python
